feat: Refactor authentication and user management
- Updated AuthController to inherit from BaseController and changed response handling to use SuccessResponse.
- Added BankAccountController for managing bank accounts with a create endpoint.
- Introduced BankAccount entity with properties for account management.
- Modified BaseModel to change CreatedBy and UpdatedBy types from int to long.
- Created IBankAccountReadRepository and IBankAccountWriteRepository interfaces for bank account data access.
- Implemented CreateBankAccountCommand and its handler for bank account creation logic.
- Enhanced CreateUserCommand to include validation attributes for required fields.
- Updated GetUserProfileQueryHandler to use long for user ID comparison.
- Removed obsolete PayExpressDbContext and related user model files.
- Updated AppDbContext to include DbSet for BankAccounts.
- Registered new repositories and services in ServiceCollectionExtensions.
- Added PinHasher service for secure PIN handling.
- Created migrations for initial database setup including Users and BankAccounts tables.

+++ Core/Interfaces/Repositories/Read/IBaseReadRepository.cs
@@ -4,7 +4,7 @@ namespace PaymentCoreServiceApi.Core.Interfaces.Repositories.Read;
 
 public interface IBaseReadRepository<TEntity> where TEntity : class
 {
-    Task<TEntity> GetByIdAsync(int id);
+    Task<TEntity> GetByIdAsync(long id);
     Task<IEnumerable<TEntity>> GetAllAsync();
     Task<(IEnumerable<TEntity> Items, int TotalCount)> GetPagedAsync(int pageIndex, int pageSize, Expression<Func<TEntity, bool>> predicate);
 }
